Etsy StatsD compatibility work and error logging
Remove percentiles for the median.  These are not handled by Etsy's
StatsD and we want to be compatible.

Log when we have a TCP connection error to Carbon and if we retry that
metric.  We want to have some idea if we start dropping metrics due to
continued errors / retries.